Worked around absence of isnan in C++98 and C89
The symbol isnan is defined neither in the C++98 nor in the C89 standards.

It is specified in the C99 standard, under the math.h header,
where it is implemented as a macro defined in the global namespace.
It is also specified by the C++11 standard -- and recommended in TR1 --
under the cmath header, where it corresponds to a set of overloaded functions
defined in the std namespace.

The most straightforward and future-proof solution is to adhere to the C++11 specification,
and therefore use the std namespace qualification for isnan.
A workaround will be required if is determined that a non-compliant compiler must be also supported.